Synchronization

+++ game/image_graphics.h
@@ -1,6 +1,6 @@
 #ifndef IMAGE_GRAPHICS_H
 #define IMAGE_GRAPHICS_H
 
-void display(unsigned int **screen);
+void display(unsigned short **screen);
 
 #endif
