+++ game/stats.h
@@ -1,12 +1,12 @@
 #ifndef STATS_H
 #define STATS_H
 
-void print_statistics(unsigned int **screen, unsigned int *statistics);
+void print_statistics(unsigned short **screen, unsigned int *statistics);
 
-void print_score(unsigned int **screen, unsigned int score);
+void print_score(unsigned short **screen, unsigned int score);
 
-void print_best_score(unsigned int **screen, unsigned int best_score);
+void print_best_score(unsigned short **screen, unsigned int best_score);
 
-void print_destroyed_lines_number(unsigned int **screen, unsigned int lines_amount);
+void print_destroyed_lines_number(unsigned short **screen, unsigned int lines_amount);
 
 #endif
