+++ game/tools/font.h
@@ -1,6 +1,6 @@
 #ifndef FONT_H
 #define FONT_H
 
-void draw_string(unsigned int **screen, int x, int y, char *str, unsigned int fg_color, unsigned int bg_color);
+void draw_string(unsigned short **screen, int x, int y, char *str, unsigned int fg_color, unsigned int bg_color);
 
 #endif
